Added the 3 Neural Network inputs:
-birdHeights[]. (list of all birds heights from floor)

-distToNextPipe. distance of birds to next Pipe. birds cant change their 
x position, so we only need to get the x position of 1 bird.

-nextPipeHeight. Like above, there will never be an isntance where 2 
active birds are in different pipe stages, so we only need 1 value.

--
Tweaked when the bird thinks it has passed a pipe, and when the new pipe 
goal is updated.
Previously the pipe goal only updated when a pipe set is deleted (bird 
was halfway to the next one when this happened). Now it is triggered 
when the bird reaches the end of 1 pipe-set.
When it did not update correctly, it also effected the update of the 
nextPipeHeight. Obviously both issues are now resolbed from the fix.
--
Added Debug lines that accurately depict the 3 Neural Network inputs.